face_model/requirements.txt
kagglehub
ultralytics
opencv-python
fastapi
uvicorn
python-multipart
pillow
pyyaml
